+++ src/formatters/rubyfmt.ts
@@ -22,15 +22,14 @@ export class RubyfmtFormatter extends Formatter {
     _isRange?: boolean,
     token?: vscode.CancellationToken,
   ): Promise<string | undefined> {
-    const options = {
-      cmd: this.getFormatterCommand(uri),
-      cwd: this.getCwd(uri),
-    };
-
-    if (!(await verifyFormatter(this.context, uri, options, this.descriptor))) {
+    if (!(await verifyFormatter(this.context, uri, this))) {
       return;
     }
 
-    return this.run(text, uri, options, token);
+    return this.run(text, uri, {}, token);
+  }
+
+  public override getFormatterCommand(scope?: vscode.ConfigurationScope): string[] {
+    return [this.context.configuration.getFormatterPath(this.id, scope)];
   }
 }