+++ src/formatters/standardrb.ts
@@ -51,12 +51,7 @@ export class StandardRbFormatter extends Formatter {
     token?: vscode.CancellationToken,
   ): Promise<string | undefined> {
     const { uri } = document;
-    const options = {
-      cmd: this.getFormatterCommand(uri),
-      cwd: this.getCwd(uri),
-    };
-
-    if (!(await verifyFormatter(this.context, uri, options, this.descriptor))) {
+    if (!(await verifyFormatter(this.context, uri, this))) {
       return;
     }
 
@@ -71,7 +66,7 @@ export class StandardRbFormatter extends Formatter {
       case 'forceSave':
         return await this.formatAfterSave(document, text, token);
       case 'tmpFile':
-        return await this.formatWithTmpFile(document, text, token);
+        return await this.formatWithTemporaryFile(document, text, token);
       default:
         throw new Error(`Unsupported StandardRB formatting mode: ${mode}`);
     }
@@ -94,7 +89,7 @@ export class StandardRbFormatter extends Formatter {
     return await this.runStandardRb(document.uri, text, true, token);
   }
 
-  private async formatWithTmpFile(
+  private async formatWithTemporaryFile(
     _document: vscode.TextDocument,
     text: string,
     token?: vscode.CancellationToken,